requirements.txt
tqdm==4.66.5
groq==0.9.0
python-dotenv==1.0.1
pymongo==4.8.0